+++ inst/testme/test-findGlobals,dfs.R
@@ -207,9 +207,20 @@ append_expr(expr, substitute = FALSE, truth = c("for"))
 expr <- quote(for (x in NULL) x)
 append_expr(expr, substitute = FALSE, truth = c("for"))
 
-
 expr <- quote(base::names(x)[1] <- 0)
 append_expr(expr, substitute = FALSE, truth = c("::", "x", "[<-"))
+
+expr <- alist(x = a)
+append_expr(expr, substitute = FALSE, truth = c("a"))
+
+expr <- function(x = a) NULL
+append_expr(expr, substitute = FALSE, truth = c("a"))
+
+expr <- function(...) list(..1, ..2)
+append_expr(expr, substitute = FALSE, truth = c("list"))
+
+expr <- function() list(..1, ..2)
+append_expr(expr, substitute = FALSE, truth = c("list", "..1", "..2"))
 
 for (kk in seq_along(exprs)) {
   message(sprintf("\n*** Expression #%d ***", kk))